refactor: InvitationButton 리팩토링 및 public 폴더 중복 이미지 파일 제거
