+++ app/mydashboard/_components/organisms/DashboardButton.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import styled from 'styled-components';
 
 import crownIcon from '@public/images/icon-crown.svg';
-import rightArrowIcon from '@public/images/icon-rignt-arrow-default.svg';
+import rightArrowIcon from '@public/images/icon-right-arrow.svg';
 import { mediaBreakpoint } from '@styles/mediaBreakpoint';
 
 // @ToDo 상하 정렬 수정해야할듯
